+++ examples/webgl/src/common/type.ts
@@ -1,0 +1,3 @@
+export interface Renderer {
+  render(lastRendered: string): string;
+}
