+++ examples/webgl/src/cube/frag.glsl
