+++ examples/webgl/src/cube/vert.glsl
