+++ examples/webgl/src/index.ts
@@ -1,7 +1,6 @@
-import CubeWorld from './cube_world';
+import CubeWorld from './world';
 
 const world = new CubeWorld('hello-world');
 
 world.attach(document.body);
-world.add('1', [1, 0, 0, 1]);
 world.render();
