+++ examples/webgl/tsconfig.json
@@ -12,6 +12,7 @@
       // Add this line to get types for streaming imports (packageOptions.source="remote"):
       //     "*": [".snowpack/types/*"]
       // More info: https://www.snowpack.dev/guides/streaming-imports
+      "@common/*": ["./src/common/*"]
     },
     "allowSyntheticDefaultImports": true,
     "importsNotUsedAsValues": "error",
